requirements.txt
sqlmodel
fastapi
wtforms
sqlalchemy
starlette-wtf
pint
pint
pyyaml
pydantic
python-multipart
minio
psycopg2
orator
